add with-expect-call2 that allows unordered calls
It also restructures how the lib works completely to make it more
extensible and clean. The new version allows unordered mock calls and
can allow for multiple calls of the same fn in any order. Currently the
lib is locked into ordered calls. Requiring ordered calls could be added
to this POC. So far :do :more, and :once are supported. Arg matching
also works as the previous one.

Need to add a post-run check to make sure the mocks are all called.
This should be as easy as checking the number of calls on each of the mocks.